src/test/java/ru/job4j/synch/SimpleBlockingQueueTest.java
package ru.job4j.synch;

import org.junit.Test;

import static org.junit.Assert.*;

public class SimpleBlockingQueueTest {
    @Test
    public void whenEmpty() throws InterruptedException {
        SimpleBlockingQueue<Integer> queue = new SimpleBlockingQueue<>();
        Thread produser = new Thread(() -> {
            for (int i = 0; i < 10; i++) {
                queue.offer(i);
            }
        });
        Thread consumer = new Thread(() -> {
            for (int i = 0; i < 10; i++) {
                queue.poll();
            }
        });
        produser.start();
        consumer.start();
        produser.join();
        consumer.join();
        assertTrue(queue.isEmpty());
    }

    @Test
    public void whenNotEmpty() throws InterruptedException {
        SimpleBlockingQueue<Integer> queue = new SimpleBlockingQueue<>();
        Thread produser = new Thread(() -> {
            for (int i = 0; i < 10; i++) {
                queue.offer(i);
            }
        });
        Thread consumer = new Thread(() -> {
            for (int i = 0; i < 9; i++) {
                queue.poll();
            }
        });
        produser.start();
        consumer.start();
        produser.join();
        consumer.join();
        assertFalse(queue.isEmpty());
    }
}